currency API fixed and updated, many small changes

+++ app/currencies.php
@@ -14,10 +14,16 @@ if($time === $lastUpdate) {
 }
 
 //or get a new JSON, save it and send it
+//uses the free API for currency exchange rates: https://fixer.io/
 else {
-	$data = file_get_contents('http://api.fixer.io/latest?base=USD');
-	file_put_contents('currencies.json', $data);
-	//counter of requests:
-	//touch('foobar');file_put_contents('foobar', (int)file_get_contents('foobar') + 1);
-	echo $data;
+	//your personal API Key is stored in this file. I suggest denying access to it (e.g. using .htaccess)
+	if(file_exists('API_key')) {
+		$key = file_get_contents('API_key');
+	
+		$data = file_get_contents('http://data.fixer.io/api/latest?access_key='.$key);
+		file_put_contents('currencies.json', $data);
+		//counter of requests, uncomment if you want it:
+		//touch('foobar');file_put_contents('foobar', (int)file_get_contents('foobar') + 1);
+		echo $data;
+	}
 }
